Encouragement
This program sends encouraging messages based on the input. If "A" is pressed, there are 5 possible messages to be outputted. If "B" is pressed, the program chooses a percentage and outputs it, with a message saying how awesome the user is today. If both "A" and "B" are pressed, the program repeats a message four times.